# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Lifio API
  description: |
    Backend API for the Lifio personal time-tracking iOS app.

    ## Auth flow (passwordless email-OTP)
    1. **Request OTP** — `POST /api/v1/auth/otp/request` with an email address. Always returns 202
       (prevents user enumeration). The server upserts the user, generates a 6-digit code,
       and sends it via email (console in dev/test, AWS SES in production).
    2. **Verify OTP** — `POST /api/v1/auth/otp/verify` with the email and 6-digit code. On success,
       returns a JWT access token (15 min TTL) and an opaque refresh token.
    3. **Refresh** — `POST /api/v1/auth/refresh` with the refresh token and `X-Device-Id`.
       Issues a new token pair within the sending device's refresh family and revokes the
       old token. Reuse of a revoked token revokes only that device's family — other
       devices' sessions keep working.
    4. **Logout** — `POST /api/v1/auth/logout` with a Bearer access token and `X-Device-Id`.
       Revokes only the calling device's refresh family; other devices' sessions remain valid.
    5. **Me** — `GET /api/v1/auth/me` with a Bearer access token. Returns the user profile.

    ## Auth flow (Sign in with Apple)
    - **Apple** — `POST /api/v1/auth/apple` with Apple's identity-token JWT (RS256). The server
      verifies it against Apple's JWKS (iss/aud/exp) and upserts a user keyed by Apple's
...
       The route is always registered; without `APPLE_CLIENT_ID` it returns 503
...

...

...

...

...
servers:
...
paths:
...
  /api/v1/entries:
    get:
      operationId: listEntries
      summary: List entries
      description: |
        Returns the authenticated user's entries, newest first, with filters and
        cursor pagination. Each entry carries its own `activity_text`, ordered
        category tags, and notes — no record is resolved at query time. Used by
        History; entries sync cross-device from the start.
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive lower bound on `started_at`.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inclusive upper bound on `started_at`.
        - name: category_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to entries tagged with this category (via `entry_categories`).
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor returned by a previous request.
        - name: modified_since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: |
            Delta pull-sync: return only entries with `updated_at` strictly
            greater than this RFC 3339 timestamp. Absent/empty = full pull.
      responses:
        "200":
          description: A page of entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createEntry
      summary: Create a time entry
      description: |
        Creates a time entry. `activity_text` is required (trimmed, non-empty,
        max 60 chars; case-sensitive identity); `category_ids` are stored on
        the entry via the `entry_categories` join (order preserved, unknown or
        non-owned ids rejected with 422). `ended_at` null/omitted = running.
        `id` is client-generated UUID v7 (idempotent on replay).
      tags:
        - Entries
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryCreate"
      responses:
        "201":
          description: Entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "200":
          description: Idempotent replay — an entry with this `id` already exists and is returned unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          $ref: "#/components/responses/InvalidBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/entries/recents:
    get:
      operationId: listRecents
      summary: List recent entry texts
      description: |
        Returns up to `limit` entries (default 6, max 20) — the newest entry
        per exact `activity_text` (`GROUP BY activity_text`, newest
        `started_at` wins per group, `id DESC` tiebreak), newest-first. Each
        item is a full entry with its own ordered category tags, so a fresh
        device can render Recents (exact text + first-position category icon)
        before its first full entry pull. Identity is byte-exact trimmed text,
        case-sensitive (`Gym` ≠ `GYM`).
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
          description: Max exact-text groups to return.
      responses:
        "200":
          description: The recents entries, newest group first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/entries/{id}:
    get:
      operationId: getEntry
      summary: Get an entry
      description: Returns a single entry with its own ordered category tags.
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateEntry
      summary: Update an entry
      description: |
        Updates an entry (last-write-wins on `updated_at`). `activity_text`,
        `notes`, and `category_ids` (full replace, order preserved) are
        editable; setting `ended_at` stops a running timer and recomputes
        `duration_seconds`.
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: The updated entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          $ref: "#/components/responses/InvalidBody"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteEntry
      summary: Delete an entry
      description: |
        Hard-deletes a single entry (F10 "delete only the current entry"). Undo is client-side:
        the deletion sits in the durable undo buffer, restorable until the app restarts (no wall-clock window).
      tags:
        - Entries
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Entry deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------- Cross-device deletions ----------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/otp/verify`, `/api/v1/auth/apple`, or `/api/v1/auth/refresh`.
        Format: `Authorization: Bearer <token>`
  responses:
...
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unauthorized
              message: Invalid or expired access token
    InvalidBody:
      description: Malformed request body (invalid JSON or extra data).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: invalid_body
              message: Invalid request body
    InternalError:
      description: An unexpected server-side failure occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: internal_error
              message: An internal error occurred
    ValidationError:
      description: Field-level validation failed; `details` maps each invalid field to a single unified message.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: validation_error
              message: Validation failed
              details:
                name: Name must be 60 characters or fewer
    NotFound:
      description: Resource not found, or it belongs to another user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: not_found
              message: Not found
    Conflict:
      description: |
        409 conflict. `conflict` — a last-write-wins stale write (`updated_at`
        older than the server's); `details` carries the server's current version.
        `category_exists` — a case-insensitive name collision on a category
        create; `details` carries the existing record's id and name.
        `duplicate_import` — an entry with the same (source, source_ref) already
        exists (duplicate import prevention).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            conflict:
              summary: Stale last-write-wins write
              value:
                error:
                  code: conflict
                  message: Outdated version; a newer record exists on the server.
                  details:
                    updated_at: "2026-07-27T10:00:00Z"
            category_exists:
              summary: Category name already exists (case-insensitive)
              value:
                error:
                  code: category_exists
                  message: A category with this name already exists.
                  details:
                    id: "019639f1-7a3b-7abc-9def-100000000002"
                    name: Sport
            duplicate_import:
              summary: Entry with the same source and source_ref already exists
              value:
                error:
                  code: duplicate_import
                  message: An entry with this source and source_ref already exists.
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Uniform error envelope used by all endpoints.
        The `details` field may contain additional context (e.g. validation errors).
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: invalid_body
            message:
              type: string
              description: Human-readable error description
              example: Invalid request body
            details:
              type: object
              description: Optional additional error context
              example: {}

    # ---------- Catalog & entries schemas ----------

    # Shared embedded tag (a category as it appears on an entry).
    CategoryTag:
      type: object
      description: A category tag embedded in an entry.
      required:
        - id
        - name
        - icon
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          $ref: "#/components/schemas/CategoryIcon"

    # Authoritative category icon catalog (category-management spec D1): the
    # closed set of SF Symbol names the relay accepts. The Go validator
    # (`internal/handlers.validIcons`) and the iOS `CatalogIcon` type mirror
    # this exact list; the contract test `TestSpec_CategoryIconEnumMatchesGo`
    # fails when they drift.
    CategoryIcon:
      type: string
      description: |
        The supported SF Symbol catalog for category icons. This enum is the
        machine-readable canonical list; the Go `validIcons` set and the iOS
        `CatalogIcon` type mirror it, and `Design/TOKENS.md` documents the
        same set. The icon picker hides catalog symbols unavailable on the
        running OS; a valid synchronized symbol that cannot render is
        displayed as `tag` without changing the stored value.
      enum:
        - clock
        - laptopcomputer
        - briefcase
        - book
        - pencil.and.ruler
        - brain.head.profile
        - figure.walk
        - figure.run
        - figure.strengthtraining
        - figure.yoga
        - figure.cycling
        - figure.swimming
        - figure.soccer
        - figure.basketball
        - figure.tennis
        - figure.gymnastics
        - figure.mindandbody
        - figure.core.training
        - dumbbell
        - bicycle
        - books
        - graduationcap
        - desktopcomputer
        - keyboard
        - gamecontroller
        - fork.knife
        - cup.and.saucer
        - bed.double
        - moon.stars
        - moon.zzz
        - film
        - music.note
        - guitar
        - camera
        - tv
        - musicalnotes
        - paintbrush
        - house
        - car.fill
        - airplane
        - cart
        - phone
        - hammer
        - heart
        - leaf
        - sparkles
        - pawprint
        - dog
        - cat
        - fish
        - bird
        - washer
        - dryer
        - dishwasher
        - refrigerator
        - sofa
        - shower
        - lamp.table
        - person.2
        - figure.and.child.holdinghands
        - stethoscope
        - pill
        - tag

    # ---------- Catalog request schemas ----------
...
    EntryCreate:
      type: object
      description: |
        Request body for creating a time entry. `activity_text` is required
        (trimmed, non-empty, max 60 chars, case-sensitive identity — `Gym` ≠
        `GYM`); `category_ids` are stored on the entry (order preserved);
        unknown ids are rejected with 422. `ended_at` omitted/null means a
        running timer. `notes` defaults to empty. `source` (default `manual`)
        and `source_ref` record provenance; a duplicate (source, source_ref)
        is rejected with 409 `duplicate_import`.
      required:
        - id
        - activity_text
        - started_at
      properties:
        id:
          type: string
          format: uuid
          description: Client-generated UUID v7.
        activity_text:
          type: string
          minLength: 1
          maxLength: 60
          description: The entry's display text (trimmed; case-sensitive identity per user — no uniqueness constraint).
        notes:
          type: string
          maxLength: 280
          description: Entry notes; defaults to empty.
        category_ids:
          type: array
          items:
            type: string
            format: uuid
          description: |
            Ordered category tags for the entry (stored via `entry_categories`).
            Duplicates and empty strings are skipped; every id must be an
            existing category of the user (else 422).
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
          description: End time; null/omitted = running timer.
        source:
          type: string
          enum: [manual, widget, siri, control, screentime, garmin, calendar, healthkit]
          description: Entry provenance; defaults to `manual` when omitted.
        source_ref:
          type: string
          nullable: true
          description: External identifier for the source; null for manual entries.
      example:
        id: "019639f1-7a3b-7abc-9def-100000000003"
        activity_text: Gym
        notes: ""
        category_ids:
          - "019639f1-7a3b-7abc-9def-100000000002"
        started_at: "2026-07-27T09:00:00Z"
        source: manual
        source_ref: null
    EntryUpdate:
      type: object
      description: |
        Request body for updating an entry. All fields optional except
        `updated_at` (last-write-wins version). `category_ids`, if present, is
        a full replace of the entry's ordered tags; omit it to leave tags
        unchanged.
      required:
        - updated_at
      properties:
        activity_text:
          type: string
          minLength: 1
          maxLength: 60
          description: Replacement display text (trimmed; case-sensitive identity).
        notes:
          type: string
          maxLength: 280
          description: Replacement notes.
        category_ids:
          type: array
          items:
            type: string
            format: uuid
          description: Replace-all ordered tag set (unknown/non-owned ids pruned, remainder kept).
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
          description: Setting this stops a running timer; `duration_seconds` is recomputed.
        updated_at:
          type: string
          format: date-time
          description: Client's version timestamp; the write applies only if newer than the server's.

    # ---------- Catalog & entries response schemas ----------
...
    Entry:
      type: object
      description: |
        A time entry. Every entry owns its `activity_text` (trimmed display
        text, case-sensitive identity — history never mutates retroactively),
        its ordered `categories` (stored via `entry_categories`), and its
        `notes`. `source` records where the entry came from (provenance);
        `source_ref` holds the external identifier for that source (e.g. a
        Screen Time callback uuid or Garmin activity id) and is null for
        manual entries.
      required:
        - id
        - activity_text
        - notes
        - started_at
        - source
        - created_at
        - updated_at
        - categories
      properties:
        id:
          type: string
          format: uuid
        activity_text:
          type: string
          maxLength: 60
          description: The entry's display text (owned by the entry; never resolved from another record).
        notes:
          type: string
          maxLength: 280
          description: Entry notes; empty when unset.
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
          description: Null = running timer.
        duration_seconds:
          type: integer
          nullable: true
          description: ended_at - started_at, in seconds; null while running.
        source:
          type: string
          enum: [manual, widget, siri, control, screentime, garmin, calendar, healthkit]
          description: |
            Entry provenance — where the entry came from. `manual` is the
            default for entries created without a source.
        source_ref:
          type: string
          nullable: true
          description: |
            External identifier for the source (e.g. Screen Time callback uuid,
            Garmin activity id, `.ics` UID). Null for manual entries. The
            server enforces uniqueness on (user_id, source, source_ref) for
            non-null source_ref, preventing duplicate imports.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryTag"
      example:
        id: "019639f1-7a3b-7abc-9def-100000000003"
        activity_text: Gym
        notes: ""
        started_at: "2026-07-27T09:00:00Z"
        ended_at: "2026-07-27T10:00:00Z"
        duration_seconds: 3600
        source: manual
        source_ref: null
        created_at: "2026-07-27T09:00:00Z"
        updated_at: "2026-07-27T10:00:00Z"
        categories:
          - id: "019639f1-7a3b-7abc-9def-100000000002"
            name: Sport
            icon: tag
    EntryListResponse:
      type: object
      description: Paginated entries list, newest first.
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
        next_cursor:
          type: string
          nullable: true
...